@@ -1,30 +1,14 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 
-
-tasks.withType(Javadoc) {
-    options.addStringOption('Xdoclint:none', '-quiet')
-    options.addStringOption('encoding', 'UTF-8')
-    options.addStringOption('charSet', 'UTF-8')
-}
-
-tasks.withType(Javadoc).all {
-    enabled = false
-}
-
 android {
-    compileSdkVersion 28
-
-
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 29
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -39,39 +23,11 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:28.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.2.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-}
-repositories {
-    mavenCentral()
+
 }
 
-
-tasks.withType(Javadoc).all { enabled = false }
-
-ext {
-    package_userOrg = 'swordsoul'
-    package_repo = 'line'
-    package_type = 'aar'
-    package_group = 'com.line'
-    package_artifact = 'swipebackhelper'
-    package_version = '1.0'
-    package_description = 'A simple swipe back helper for android developer'
-
-    siteUrl = 'https://github.com/microtears/SwipeBackHelper'
-    gitUrl = 'https://github.com/microtears/SwipeBackHelper.git'
-
-    developerId = 'line'
-    developerName = 'line'
-    developerEmail = 'microtears@gmail.com'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = 'Apache-2.0'
-}
-
-//apply from: 'https://raw.githubusercontent.com/Jude95/JCenter/v2/bintray.gradle'
-apply from: 'bintray_upload.gradle'
+apply from: 'publish.gradle'
