@@ -1,16 +1,15 @@
-package com.messy.swipebackdemo
+package com.microtears.swipebackdemo
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.MotionEvent
-import com.messy.swipebackhelper.SwipeBackHelper
+import androidx.appcompat.app.AppCompatActivity
+import com.microtears.swipebackhelper.SwipeBackHelper
 import kotlinx.android.synthetic.main.activity_second.*
 
 class SecondActivity : AppCompatActivity() {
 
-    private val swipeBackHelper = SwipeBackHelper()
-    //private val swipeBackHelper = SwipeBackHelper().apply { isFullScreen = true/*开启全面屏适配*/ }
+    private val swipeBackHelper = SwipeBackHelper().apply { isFullScreen = true/*开启全面屏适配*/ }
 
     override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
         return if (!swipeBackHelper.progressTouchEvent(ev))
