@@ -1,12 +1,12 @@
-package com.messy.swipebackdemo
+package com.microtears.swipebackdemo
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.MotionEvent
-import com.messy.swipebackhelper.SwipeBackHelper
+import androidx.appcompat.app.AppCompatActivity
+import com.microtears.swipebackhelper.SwipeBackHelper
 
 class ThirdActivity : AppCompatActivity() {
-    private val swipeBackHelper = SwipeBackHelper()
+    private val swipeBackHelper = SwipeBackHelper().apply { isFullScreen = true/*开启全面屏适配*/ }
 
     override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
         return if (!swipeBackHelper.progressTouchEvent(ev))
